requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
Django==3.0.6
django-taggit==1.2.0
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
